+++ sdk/servicebus/azure-servicebus/azure/servicebus/__init__.py
@@ -13,7 +13,7 @@ from ._servicebus_sender import ServiceBusSender
 from ._servicebus_receiver import ServiceBusReceiver
 from ._servicebus_session import ServiceBusSession
 from ._common.message import ServiceBusMessage, ServiceBusMessageBatch, ServiceBusReceivedMessage
-from ._common.constants import ServiceBusReceiveMode, SubQueue, NEXT_AVAILABLE_SESSION
+from ._common.constants import ServiceBusReceiveMode, ServiceBusSubQueue, NEXT_AVAILABLE_SESSION
 from ._common.auto_lock_renewer import AutoLockRenewer
 from ._common._connection_string_parser import (
     parse_connection_string,
@@ -27,7 +27,7 @@ __all__ = [
     'ServiceBusMessageBatch',
     'ServiceBusReceivedMessage',
     'NEXT_AVAILABLE_SESSION',
-    'SubQueue',
+    'ServiceBusSubQueue',
     'ServiceBusReceiveMode',
     'ServiceBusClient',
     'ServiceBusReceiver',