+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_client_async.py
@@ -13,7 +13,7 @@ from ._servicebus_sender_async import ServiceBusSender
 from ._servicebus_receiver_async import ServiceBusReceiver
 from .._common._configuration import Configuration
 from .._common.utils import generate_dead_letter_entity_name
-from .._common.constants import SubQueue
+from .._common.constants import ServiceBusSubQueue
 from ._async_utils import create_authentication
 
 if TYPE_CHECKING:
@@ -205,7 +205,8 @@ class ServiceBusClient(object):
          sessionful queue, otherwise it must be None. In order to receive messages from the next available
          session, set this to ~azure.servicebus.NEXT_AVAILABLE_SESSION.
         :paramtype session_id: Union[str, ~azure.servicebus.NEXT_AVAILABLE_SESSION]
-        :keyword Optional[SubQueue] sub_queue: If specified, the subqueue this receiver will connect to.
+        :keyword Optional[Union[ServiceBusSubQueue, str]] sub_queue: If specified, the subqueue this receiver will
+         connect to.
          This includes the DeadLetter and TransferDeadLetter queues, holds messages that can't be delivered to any
          receiver or messages that can't be processed.  The default is None, meaning connect to the primary queue.
         :keyword receive_mode: The mode with which messages will be retrieved from the entity. The two options
@@ -246,11 +247,14 @@ class ServiceBusClient(object):
                 "To connect to the sub queue of a sessionful queue, "
                 "please set sub_queue only as sub_queue does not support session."
             )
-        if sub_queue and sub_queue in SubQueue:
+        try:
             queue_name = generate_dead_letter_entity_name(
                 queue_name=queue_name,
-                transfer_deadletter=(sub_queue == SubQueue.TransferDeadLetter)
+                transfer_deadletter=(ServiceBusSubQueue(sub_queue) == ServiceBusSubQueue.TransferDeadLetter)
             )
+        except ValueError:
+            if sub_queue: # If we got here and sub_queue is defined, it's an incorrect value or something unrelated.
+                raise
         handler = ServiceBusReceiver(
             fully_qualified_namespace=self.fully_qualified_namespace,
             entity_name=queue_name,
@@ -311,7 +315,8 @@ class ServiceBusClient(object):
          sessionful subscription, otherwise it must be None. In order to receive messages from the next available
          session, set this to ~azure.servicebus.NEXT_AVAILABLE_SESSION.
         :paramtype session_id: Union[str, ~azure.servicebus.NEXT_AVAILABLE_SESSION]
-        :keyword Optional[SubQueue] sub_queue: If specified, the subqueue this receiver will connect to.
+        :keyword Optional[Union[ServiceBusSubQueue, str]] sub_queue: If specified, the subqueue this receiver will
+         connect to.
          This includes the DeadLetter and TransferDeadLetter queues, holds messages that can't be delivered to any
          receiver or messages that can't be processed.  The default is None, meaning connect to the primary queue.
         :keyword receive_mode: The mode with which messages will be retrieved from the entity. The two options
@@ -353,11 +358,11 @@ class ServiceBusClient(object):
                 "To connect to the sub queue of a sessionful subscription, "
                 "please set sub_queue only as sub_queue is not sessionful."
             )
-        if sub_queue and sub_queue in SubQueue:
+        try:
             entity_name = generate_dead_letter_entity_name(
                 topic_name=topic_name,
                 subscription_name=subscription_name,
-                transfer_deadletter=(sub_queue == SubQueue.TransferDeadLetter)
+                transfer_deadletter=(ServiceBusSubQueue(sub_queue) == ServiceBusSubQueue.TransferDeadLetter)
             )
             handler = ServiceBusReceiver(
                 fully_qualified_namespace=self.fully_qualified_namespace,
@@ -373,7 +378,9 @@ class ServiceBusClient(object):
                 retry_backoff_max=self._config.retry_backoff_max,
                 **kwargs
             )
-        else:
+        except ValueError:
+            if sub_queue: # If we got here and sub_queue is defined, it's an incorrect value or something unrelated.
+                raise
             handler = ServiceBusReceiver(
                 fully_qualified_namespace=self.fully_qualified_namespace,
                 topic_name=topic_name,