+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_subscriptions_async.py
@@ -15,7 +15,7 @@ from azure.servicebus import ServiceBusMessage, ServiceBusReceiveMode
 from azure.servicebus.aio import ServiceBusClient
 from azure.servicebus.aio._base_handler_async import ServiceBusSharedKeyCredential
 from azure.servicebus.exceptions import ServiceBusError
-from azure.servicebus._common.constants import SubQueue
+from azure.servicebus._common.constants import ServiceBusSubQueue
 
 from devtools_testutils import AzureMgmtTestCase, RandomNameResourceGroupPreparer, CachedResourceGroupPreparer
 from servicebus_preparer import (
@@ -157,7 +157,7 @@ class ServiceBusSubscriptionAsyncTests(AzureMgmtTestCase):
             async with sb_client.get_subscription_receiver(
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
-                sub_queue = SubQueue.DeadLetter,
+                sub_queue = ServiceBusSubQueue.DeadLetter,
                 max_wait_time=5,
                 receive_mode=ServiceBusReceiveMode.PeekLock
             ) as dl_receiver: